Implement md2 model animation (VAT) #117: vertex data ready
Implemented the vertex positional data conversion from md2 format.
Turned out to be easier than expected - instead of reindexing all the vertices
in the frames positions, the index in the VAT is passed as the vertex attribute
 (along with the texture coordinates).
Not only this helps to avoid duplication of data, it also makes the code simpler.

Added a test.